+++ crates/ai/src/ai.rs
@@ -2,3 +2,4 @@ pub mod completion;
 pub mod embedding;
 pub mod models;
 pub mod prompts;
+pub mod providers;
