+++ crates/ai/src/providers/open_ai/mod.rs
@@ -1,0 +1,2 @@
+pub mod model;
+pub use model::OpenAILanguageModel;
